create history page component
